Move prop types to types.ts, separate AttachmentPreview lookups
